Upgrade pillow
