+++ .flake8
@@ -1,0 +1,3 @@
+[flake8]
+exclude=.venv,venv
+max-line-length=120
