--- directory_validators/constants/__init__.py
@@ -1,4 +1,0 @@
-NO_EXPORT_INTENTION_ERROR_LABEL = (
-    'You must intend to export to use this service.'
-)
-ALLOWED_IMAGE_FORMATS = ('PNG', 'JPG', 'JPEG')
