--- directory_validators/constants/disposable_email_domains.py
@@ -1,878 +1,0 @@
-disposable_domains = [
-    '0-mail.com',
-    '0815.ru',
-    '0clickemail.com',
-    '0sg.net',
-    '0wnd.net',
-    '0wnd.org',
-    '10minutemail.com',
-    '10minutemail.com ',
-    '12hourmail.com',
-    '1chuan.com',
-    '1zhuan.com',
-    '20minutemail.com',
-    '21cn.com',
-    '2prong.com',
-    '2prong.com ',
-    '30minutemail.com',
-    '30minutesmail.com',
-    '3126.com',
-    '3d-painting.com',
-    '4warding.com',
-    '4warding.com ',
-    '4warding.net',
-    '4warding.org',
-    '50e.info',
-    '60minutemail.com',
-    '675hosting.com ',
-    '675hosting.net ',
-    '675hosting.org ',
-    '6url.com',
-    '6url.com ',
-    '75hosting.com ',
-    '75hosting.net ',
-    '75hosting.org ',
-    '9ox.net',
-    'BeefMilk.com',
-    'DingBone.com',
-    'FudgeRub.com',
-    'LookUgly.com',
-    'MailScrap.com',
-    'SmellFear.com',
-    'TempEmail.net',
-    'a-bc.net',
-    'abcdefghijklmnopqrstuvwxyzabcdefghijklmnopqrstuvwxyzabcdefghijk.com',
-    'abwesend.de',
-    'addcom.de',
-    'afrobacon.com ',
-    'agnitumhost.net',
-    'ajaxapp.net ',
-    'alpenjodel.de',
-    'alphafrau.de',
-    'amilegit.com',
-    'amiri.net ',
-    'amiriindustries.com ',
-    'amorki.pl',
-    'anonbox.net',
-    'anonymbox.com',
-    'antichef.com',
-    'antichef.net',
-    'antispam.de',
-    'antispam24.de',
-    'autosfromus.com',
-    'baldmama.de',
-    'baldpapa.de',
-    'ballyfinance.com',
-    'baxomale.ht.cx',
-    'beefmilk.com',
-    'betriebsdirektor.de',
-    'bigmir.net',
-    'bin-wieder-da.de',
-    'binkmail.com',
-    'bio-muesli.info',
-    'bio-muesli.net',
-    'bleib-bei-mir.de',
-    'blockfilter.com',
-    'bluebottle.com',
-    'bobmail.info',
-    'bodhi.lawlita.com',
-    'bofthew.com',
-    'bonbon.net',
-    'brefmail.com',
-    'briefemail.com',
-    'brokenvalve.com',
-    'brokenvalve.org',
-    'bsnow.net',
-    'bspamfree.org',
-    'buerotiger.de',
-    'bugmenot.com',
-    'bugmenot.com ',
-    'bumpymail.com',
-    'bumpymail.com ',
-    'buy-24h.net.ru',
-    'cashette.com',
-    'casualdx.com',
-    'center-mail.de',
-    'centermail.at',
-    'centermail.ch',
-    'centermail.com',
-    'centermail.com ',
-    'centermail.de',
-    'centermail.info',
-    'centermail.net',
-    'centermail.net ',
-    'cghost.s-a-d.de',
-    'chogmail.com',
-    'choicemail1.com ',
-    'chongsoft.org',
-    'cool.fr.nf',
-    'coole-files.de',
-    'correo.blogos.net',
-    'cosmorph.com',
-    'courriel.fr.nf',
-    'courrieltemporaire.com',
-    'cubiclink.com',
-    'curryworld.de',
-    'cust.in',
-    'cyber-matrix.com',
-    'dacoolest.com',
-    'dandikmail.com',
-    'dating4best.net',
-    'dayrep.com',
-    'deadaddress.com',
-    'deadspam.com',
-    'deadspam.com ',
-    'deagot.com',
-    'despam.it',
-    'despam.it ',
-    'despammed.com',
-    'despammed.com ',
-    'devnullmail.com',
-    'dfgh.net',
-    'die-besten-bilder.de',
-    'die-genossen.de',
-    'die-optimisten.de',
-    'die-optimisten.net',
-    'dieMailbox.de',
-    'digital-filestore.de',
-    'digitalsanctuary.com',
-    'directbox.com',
-    'discardmail.*',
-    'discardmail.com',
-    'discardmail.com ',
-    'discardmail.de',
-    'discardmail.de ',
-    'discartmail.com',
-    'disposableaddress.com',
-    'disposeamail.com',
-    'disposeamail.com ',
-    'disposemail.com',
-    'dispostable.com',
-    'dm.w3internet.co.uk',
-    'example.com',
-    'docmail.cz',
-    'dodgeit.com',
-    'dodgeit.com ',
-    'dodgit.com',
-    'dodgit.org',
-    'dogit.com',
-    'donemail.ru',
-    'dontreg.com',
-    'dontreg.com ',
-    'dontsendmespam.de',
-    'dontsendmespam.de ',
-    'dontsentmespam.de',
-    'download-privat.de',
-    'dump-email.info',
-    'dumpandjunk.com',
-    'dumpandjunk.com ',
-    'dumpmail.com',
-    'dumpmail.de',
-    'dumpmail.de ',
-    'dumpyemail.com',
-    'dyndns.org',
-    'e-mail.com',
-    'e-mail.org',
-    'e4ward.com',
-    'e4ward.com ',
-    'eintagsmail.de',
-    'email.org',
-    'email4u.info',
-    'email60.com',
-    'emaildienst.de',
-    'emaildienst.de ',
-    'emailias.com',
-    'emailias.com ',
-    'emailigo.de',
-    'emailinfive.com',
-    'emailmiser.com',
-    'emailmiser.com ',
-    'emailsensei.com',
-    'emailtaxi.de',
-    'emailtemporanea.net',
-    'emailtemporario.com.br',
-    'emailto.de',
-    'emailto.de ',
-    'emailwarden.com',
-    'emailx.at.hm',
-    'emailxfer.com ',
-    'emz.net ',
-    'enterto.com ',
-    'ephemail.net',
-    'etranquil.com ',
-    'etranquil.net ',
-    'etranquil.org ',
-    'example.com',
-    'explodemail.com',
-    'fahr-zur-hoelle.org',
-    'fakeinbox.com',
-    'fakeinformation.com',
-    'fakeinformation.com ',
-    'falseaddress.com',
-    'fantasymail.de',
-    'farifluset.mailexpire.com',
-    'fastacura.com',
-    'fastacura.com ',
-    'fastchevy.com',
-    'fastchevy.com ',
-    'fastchrysler.com',
-    'fastchrysler.com ',
-    'fastkawasaki.com',
-    'fastkawasaki.com ',
-    'fastmazda.com',
-    'fastmazda.com ',
-    'fastmitsubishi.com',
-    'fastmitsubishi.com ',
-    'fastnissan.com',
-    'fastnissan.com ',
-    'fastsubaru.com',
-    'fastsubaru.com ',
-    'fastsuzuki.com',
-    'fastsuzuki.com ',
-    'fasttoyota.com',
-    'fasttoyota.com ',
-    'fastyamaha.com',
-    'fastyamaha.com ',
-    'feinripptraeger.de',
-    'fettabernett.de',
-    'filzmail.com',
-    'fishfuse.com',
-    'fizmail.com',
-    'forgetmail.com',
-    'fr33mail.info',
-    'frapmail.com',
-    'freemeilaadressforall.net',
-    'freudenkinder.de',
-    'fromru.com',
-    'front14.org',
-    'front14.org ',
-    'fuckingduh.com ',
-    'fux0ringduh.com ',
-    'garliclife.com',
-    'gawab.com',
-    'gentlemansclub.de',
-    'get1mail.com',
-    'get2mail.fr',
-    'getonemail.com',
-    'getonemail.com ',
-    'getonemail.net',
-    'ghosttexter.de',
-    'ghosttexter.de ',
-    'girlsundertheinfluence.com',
-    'gishpuppy.com',
-    'gishpuppy.com ',
-    'gold-profits.info',
-    'goldtoolbox.com',
-    'golfilla.info',
-    'gowikibooks.com ',
-    'gowikicampus.com ',
-    'gowikicars.com ',
-    'gowikifilms.com ',
-    'gowikigames.com ',
-    'gowikimusic.com ',
-    'gowikinetwork.com ',
-    'gowikitravel.com ',
-    'gowikitv.com ',
-    'great-host.in',
-    'greensloth.com',
-    'greensloth.com ',
-    'gsrv.co.uk',
-    'guerillamail.biz',
-    'guerillamail.com',
-    'guerillamail.net',
-    'guerillamail.org',
-    'guerrillamail.biz',
-    'guerrillamail.com',
-    'guerrillamail.com ',
-    'guerrillamail.de',
-    'guerrillamail.info',
-    'guerrillamail.net ',
-    'guerrillamail.org',
-    'guerrillamailblock.com',
-    'h.mintemail.com',
-    'h8s.org',
-    'h8s.org ',
-    'hab-verschlafen.de',
-    'habmalnefrage.de',
-    'haltospam.com',
-    'haltospam.com ',
-    'hatespam.org',
-    'hatespam.org ',
-    'herr-der-mails.de',
-    'hidemail.de',
-    'hidemail.de ',
-    'hmamail.com',
-    'hochsitze.com',
-    'home.de',
-    'hotpop.com',
-    'hulapla.de',
-    'hush.com',
-    'hushmail.com',
-    'ich-bin-verrueckt-nach-dir.de',
-    'ich-will-net.de',
-    'ieatspam.eu',
-    'ieatspam.info',
-    'ihateyoualot.info',
-    'iheartspam.org ',
-    'imails.info',
-    'imstations.com',
-    'inbox2.info',
-    'inboxclean.com',
-    'inboxclean.org',
-    'incognitomail.com',
-    'incognitomail.net',
-    'incognitomail.org',
-    'inerted.com',
-    'inmail24.com',
-    'insorg-mail.info',
-    'ipoo.org',
-    'ipoo.org ',
-    'irish2me.com',
-    'ist-allein.info',
-    'ist-einmalig.de',
-    'ist-ganz-allein.de',
-    'ist-willig.de',
-    'iwi.net ',
-    'izmail.net',
-    'jetable.com',
-    'jetable.com ',
-    'jetable.de',
-    'jetable.fr.nf',
-    'jetable.net',
-    'jetable.net ',
-    'jetable.org',
-    'jetable.org ',
-    'jetfix.ee',
-    'jetzt-bin-ich-dran.com',
-    'jn-club.de',
-    'jnxjn.com',
-    'junk1e.com',
-    'junkmail.com',
-    'kaffeeschluerfer.com',
-    'kaffeeschluerfer.de',
-    'kasmail.com',
-    'kasmail.com ',
-    'kaspop.com',
-    'keepmymail.com',
-    'killmail.com',
-    'killmail.com ',
-    'killmail.net',
-    'killmail.net ',
-    'kinglibrary.net',
-    'kir.ch.tc',
-    'klassmaster.com',
-    'klassmaster.com ',
-    'klassmaster.net',
-    'klassmaster.net ',
-    'klzlk.com',
-    'kommespaeter.de',
-    'krim.ws',
-    'kuh.mu',
-    'kulturbetrieb.info',
-    'kurzepost.de',
-    'lass-es-geschehen.de',
-    'lhsdv.com',
-    'liebt-dich.info',
-    'lifebyfood.com',
-    'link2mail.net',
-    'link2mail.net ',
-    'listomail.com',
-    'litedrop.com',
-    'lol.ovpn.to',
-    'lookugly.com',
-    'lopl.co.cc',
-    'lortemail.dk',
-    'lortemail.dk ',
-    'loveyouforever.de',
-    'lr78.com',
-    'm4ilweb.info',
-    'maboard.com',
-    'maennerversteherin.com',
-    'maennerversteherin.de',
-    'mail-temporaire.fr',
-    'mail.htl22.at',
-    'mail.mezimages.net',
-    'mail.misterpinball.de',
-    'mail.svenz.eu',
-    'mail15.com',
-    'mail2rss.org',
-    'mail2rss.org ',
-    'mail333.com',
-    'mail333.com ',
-    'mail4days.com',
-    'mail4trash.com',
-    'mail4u.info',
-    'mailbidon.com',
-    'mailblocks.com',
-    'mailblocks.com ',
-    'mailbucket.org',
-    'mailcatch.*',
-    'mailcatch.com',
-    'maileater.com',
-    'maileater.com ',
-    'mailexpire.com',
-    'mailexpire.com',
-    'mailexpire.com ',
-    'mailfreeonline.com',
-    'mailfreeonline.com ',
-    'mailin8r.com',
-    'mailin8r.com ',
-    'mailinater.com',
-    'mailinater.com ',
-    'mailinator.com',
-    'mailinator.com ',
-    'mailinator.net',
-    'mailinator2.com',
-    'mailinator2.com ',
-    'mailinblack.com',
-    'mailincubator.com',
-    'mailme.ir',
-    'mailme.lv',
-    'mailmetrash.com',
-    'mailmoat.com',
-    'mailmoat.com ',
-    'mailnator.com',
-    'mailnesia.com',
-    'mailnull.com',
-    'mailnull.com ',
-    'mailquack.com',
-    'mailshell.com',
-    'mailshell.com ',
-    'mailsiphon.com',
-    'mailsiphon.com ',
-    'mailslite.com',
-    'mailtrash.net',
-    'mailueberfall.de',
-    'mailzilla.com',
-    'mailzilla.com ',
-    'mailzilla.org',
-    'makemetheking.com',
-    'mamber.net',
-    'mbx.cc',
-    'mega.zik.dj',
-    'meine-dateien.info',
-    'meine-diashow.de',
-    'meine-fotos.info',
-    'meine-urlaubsfotos.de',
-    'meinspamschutz.de',
-    'meinspamschutz.de ',
-    'meltmail.com',
-    'messagebeamer.de',
-    'messagebeamer.de ',
-    'metaping.com',
-    'mierdamail.com',
-    'mintemail.com',
-    'mintemail.com ',
-    'mns.ru',
-    'moncourrier.fr.nf',
-    'monemail.fr.nf',
-    'monmail.fr.nf',
-    'msa.minsmail.com',
-    'msh.mailslite.com',
-    'mt2009.com',
-    'mufmail.com',
-    'muskelshirt.de',
-    'mx0.wwwnew.eu',
-    'my-mail.ch',
-    'myadult.info',
-    'mycleaninbox.net',
-    'mymail-in.net',
-    'mypartyclip.de',
-    'myphantomemail.com',
-    'myspaceinc.com ',
-    'myspaceinc.net ',
-    'myspaceinc.org ',
-    'myspacepimpedup.com ',
-    'myspamless.com ',
-    'mytempemail.com',
-    'mytop-in.net',
-    'mytrashmail.com',
-    'mytrashmail.com ',
-    'mytrashmail.compookmail.com',
-    'neomailbox.com ',
-    'nepwk.com',
-    'nervmich.net',
-    'nervmich.net ',
-    'nervtmich.net',
-    'nervtmich.net ',
-    'netmails.com',
-    'netmails.com ',
-    'netmails.net',
-    'netmails.net ',
-    'netterchef.de',
-    'netzidiot.de',
-    'netzidiot.de ',
-    'neue-dateien.de',
-    'neverbox.com',
-    'no-spam.ws',
-    'nobulk.com',
-    'nobulk.com ',
-    'noclickemail.com',
-    'noclickemail.com ',
-    'nogmailspam.info',
-    'nomail.xl.cx',
-    'nomail2me.com',
-    'nomorespamemails.com',
-    'nospam.ze.tc',
-    'nospam4.us',
-    'nospamfor.us',
-    'nospamfor.us ',
-    'nospammail.net',
-    'nospamthanks.info',
-    'notmailinator.com',
-    'notsharingmy.info',
-    'nowmymail.com',
-    'nullbox.info',
-    'nur-fuer-spam.de',
-    'nurfuerspam.de',
-    'nurfuerspam.de ',
-    'nwldx.com',
-    'nybella.com',
-    'objectmail.com',
-    'obobbo.com',
-    'office-dateien.de',
-    'oikrach.com',
-    'oneoffemail.com',
-    'oneoffemail.com ',
-    'onewaymail.com',
-    'oopi.org',
-    'oopi.org ',
-    'orangatango.com',
-    'ordinaryamerican.net',
-    'otherinbox.com',
-    'ourklips.com ',
-    'outlawspam.com ',
-    'ovpn.to',
-    'owlpic.com',
-    'pancakemail.com ',
-    'partybombe.de',
-    'partyheld.de',
-    'phreaker.net',
-    'pimpedupmyspace.com ',
-    'pisem.net',
-    'pjjkp.com',
-    'pleasedontsendmespam.de',
-    'politikerclub.de',
-    'polizisten-duzer.de',
-    'poofy.org',
-    'poofy.org ',
-    'pookmail.com',
-    'pookmail.com ',
-    'pornobilder-mal-gratis.com',
-    'portsaid.cc',
-    'postfach.cc',
-    'privacy.net',
-    'privacy.net ',
-    'privy-mail.com',
-    'proxymail.eu',
-    'prtnx.com',
-    'prydirect.info',
-    'pryworld.info',
-    'public-files.de',
-    'punkass.com',
-    'punkass.com ',
-    'put2.net',
-    'putthisinyourspamdatabase.com',
-    'qq.com',
-    'quantentunnel.de',
-    'quickinbox.com',
-    'qv7.info',
-    'ralib.com',
-    'raubtierbaendiger.de',
-    'rcpt.at',
-    'recode.me',
-    'record.me',
-    'recursor.net',
-    'regbypass.com',
-    'regbypass.comsafe-mail.net',
-    'rejectmail.com',
-    'rejectmail.com ',
-    'rklips.com ',
-    'rootprompt.org',
-    'rppkn.com',
-    'rtrtr.com',
-    's0ny.net',
-    'saeuferleber.de',
-    'safe-mail.net',
-    'safersignup.de',
-    'safersignup.de ',
-    'safetymail.info',
-    'safetypost.de',
-    'sags-per-mail.de',
-    'sandelf.de',
-    'satka.net',
-    'saynotospams.com',
-    'schmusemail.de',
-    'schreib-doch-mal-wieder.de',
-    'selfdestructingmail.com',
-    'sendspamhere.com',
-    'senseless-entertainment.com',
-    'shared-files.de',
-    'sharklasers.com',
-    'shieldedmail.com',
-    'shiftmail.com',
-    'shinedyoureyes.com',
-    'shitmail.me',
-    'shortmail.net',
-    'shortmail.net ',
-    'sibmail.com',
-    'sibmail.com ',
-    'siria.cc',
-    'skeefmail.com',
-    'skeefmail.net',
-    'slaskpost.se',
-    'slaskpost.se ',
-    'slopsbox.com',
-    'smellfear.com',
-    'sms.at',
-    'snakemail.com',
-    'sneakemail.com',
-    'sneakemail.com ',
-    'sofimail.com',
-    'sofort-mail.de',
-    'sofort-mail.de ',
-    'sofortmail.de',
-    'sogetthis.com',
-    'sogetthis.com ',
-    'sonnenkinder.org',
-    'soodonims.com',
-    'spam.la',
-    'spam.la ',
-    'spam.su',
-    'spamavert.com',
-    'spamavert.com ',
-    'spambob.com',
-    'spambob.com ',
-    'spambob.net',
-    'spambob.net ',
-    'spambob.org',
-    'spambob.org ',
-    'spambog.*',
-    'spambog.com',
-    'spambog.de',
-    'spambog.ru',
-    'spambox.info',
-    'spambox.irishspringrealty.com',
-    'spambox.us',
-    'spambox.us ',
-    'spamcannon.com',
-    'spamcannon.net',
-    'spamcero.com',
-    'spamcon.org',
-    'spamcon.org ',
-    'spamcorptastic.com',
-    'spamcowboy.com',
-    'spamcowboy.net',
-    'spamcowboy.org',
-    'spamday.com',
-    'spamday.com ',
-    'spameater.com',
-    'spameater.org',
-    'spamex.com',
-    'spamex.com ',
-    'spamfree.eu',
-    'spamfree24.com',
-    'spamfree24.com ',
-    'spamfree24.de',
-    'spamfree24.eu',
-    'spamfree24.info',
-    'spamfree24.net',
-    'spamfree24.net ',
-    'spamfree24.org',
-    'spamfree24.org ',
-    'spamgourmet.com',
-    'spamgourmet.com ',
-    'spamgourmet.net',
-    'spamgourmet.org',
-    'spamgrube.net',
-    'spamherelots.com',
-    'spamhereplease.com',
-    'spamhole.com',
-    'spamhole.com ',
-    'spamify.com',
-    'spamify.com ',
-    'spaminator.de',
-    'spaminator.de ',
-    'spamkill.info',
-    'spaml.com',
-    'spaml.com ',
-    'spaml.de',
-    'spammote.com',
-    'spammote.com ',
-    'spammotel.com',
-    'spammotel.com ',
-    'spammuffel.de',
-    'spamobox.com',
-    'spamoff.de',
-    'spamoff.de ',
-    'spamreturn.com',
-    'spamslicer.com ',
-    'spamspot.com',
-    'spamthis.co.uk',
-    'spamthisplease.com',
-    'spamtrail.com',
-    'spamtrail.com ',
-    'speed.1s.fr',
-    'sperke.net',
-    'sriaus.com',
-    'streber24.de',
-    'super-auswahl.de',
-    'supergreatmail.com',
-    'supermailer.jp',
-    'suremail.info',
-    'sweetville.net',
-    'tagesmail.eu',
-    'teewars.org',
-    'teleworm.com',
-    'temp-mail.com',
-    'temp-mail.org',
-    'tempail.com',
-    'tempalias.com',
-    'tempe-mail.com',
-    'tempemail.biz',
-    'tempemail.com',
-    'tempemail.net',
-    'tempemail.net ',
-    'tempinbox.co.uk',
-    'tempinbox.com',
-    'tempinbox.com ',
-    'tempmail.it',
-    'tempmail2.com',
-    'tempomail.fr',
-    'temporarily.de',
-    'temporarily.de ',
-    'temporarioemail.com.br',
-    'temporaryemail.net',
-    'temporaryforwarding.com',
-    'temporaryforwarding.com ',
-    'temporaryinbox.com',
-    'temporaryinbox.com ',
-    'terminverpennt.de',
-    'test.com',
-    'test.de',
-    'thanksnospam.info',
-    'thankyou2010.com',
-    'thepryam.info',
-    'thisisnotmyrealemail.com',
-    'throwawayemailaddress.com',
-    'tilien.com',
-    'tmailinator.com',
-    'topmail-files.de',
-    'tortenboxer.de',
-    'totalmail.de',
-    'tradermail.info',
-    'trash-amil.com',
-    'trash-mail.at',
-    'trash-mail.com',
-    'trash-mail.de',
-    'trash-mail.de ',
-    'trash2009.com',
-    'trashbox.eu',
-    'trashdevil.com',
-    'trashdevil.de',
-    'trashemail.de',
-    'trashmail.at',
-    'trashmail.com',
-    'trashmail.com ',
-    'trashmail.de',
-    'trashmail.de ',
-    'trashmail.me',
-    'trashmail.net',
-    'trashmail.net ',
-    'trashmail.org',
-    'trashmail.org ',
-    'trashmail.ws',
-    'trashmailer.com',
-    'trashymail.com',
-    'trashymail.net',
-    'trillianpro.com',
-    'trimix.cn',
-    'turboprinz.de',
-    'turboprinzessin.de',
-    'turual.com ',
-    'twinmail.de',
-    'twinmail.de ',
-    'tyldd.com',
-    'uggsrock.com',
-    'uk2.net',
-    'ukr.net',
-    'unterderbruecke.de',
-    'upliftnow.com ',
-    'uplipht.com ',
-    'venompen.com ',
-    'verlass-mich-nicht.de',
-    'veryrealemail.com',
-    'viditag.com ',
-    'viewcastmedia.com ',
-    'viewcastmedia.net ',
-    'viewcastmedia.org ',
-    'vinbazar.com',
-    'vollbio.de',
-    'volloeko.de',
-    'vorsicht-bissig.de',
-    'vorsicht-scharf.de',
-    'walala.org',
-    'war-im-urlaub.de',
-    'wbb3.de',
-    'webm4il.info',
-    'webmail4u.eu',
-    'wegwerfadresse.de',
-    'wegwerfadresse.de ',
-    'wegwerfemail.com',
-    'wegwerfemail.de',
-    'wegwerfmail.de',
-    'wegwerfmail.net',
-    'wegwerfmail.org',
-    'weibsvolk.de',
-    'weibsvolk.org',
-    'weinenvorglueck.de',
-    'wetrainbayarea.com ',
-    'wetrainbayarea.org ',
-    'wh4f.org',
-    'wh4f.org ',
-    'whopy.com',
-    'whyspam.me',
-    'will-hier-weg.de',
-    'willhackforfood.biz',
-    'willselfdestruct.com',
-    'willselfdestruct.com ',
-    'winemaven.info',
-    'wir-haben-nachwuchs.de',
-    'wir-sind-cool.org',
-    'wirsindcool.de',
-    'wolke7.net',
-    'women-at-work.org',
-    'wormseo.cn',
-    'wronghead.com',
-    'wuzup.net',
-    'wuzup.net ',
-    'wuzupmail.net',
-    'wwwnew.eu ',
-    'xagloo.com',
-    'xemaps.com ',
-    'xents.com',
-    'xents.com ',
-    'xmail.com',
-    'xmaily.com',
-    'xmaily.com ',
-    'xoxy.net',
-    'xsecurity.org',
-    'yep.it ',
-    'yesey.net',
-    'yogamaven.com',
-    'yopmail.com',
-    'yopmail.com ',
-    'yopmail.fr',
-    'yopmail.net',
-    'yopweb.com',
-    'youmailr.com',
-    'ystea.org',
-    'yuurok.com',
-    'yzbid.com',
-    'zehnminutenmail.de',
-    'zippymail.info',
-    'zoemail.com',
-    'zoemail.net',
-    'zoemail.org',
-    'zweb.in',
-]
