+++ setup.py
@@ -6,7 +6,7 @@ from setuptools import setup, find_packages
 
 setup(
     name='directory_validators',
-    version='5.2.0',
+    version='6.0.0',
     url='https://github.com/uktrade/directory-validators',
     license='MIT',
     author='Department for International Trade',
@@ -18,8 +18,7 @@ setup(
     install_requires=[
         'django>=1.11.22,<3.0a1',
         'olefile==0.44',
-        'phonenumberslite==7.7.4',
-        'pillow==4.2.1',
+        'pillow==6.2.1',
         'pytz==2017.2',
         'urllib3>=1.24.2<2.0.0',
     ],
